test: 자동차들을 위한 테스트 분리
